requirement.txt
amqp==2.6.1
astroid==2.4.2
backcall==0.2.0
billiard==3.5.0.5
celery==4.2.2
certifi==2020.12.5
chardet==3.0.4
cPython==0.0.6
decorator==4.4.2
Django==2.2.17
djangorestframework==3.12.2
httpie==2.3.0
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
kombu==4.3.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.3.13
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.3
pylint==2.6.0
pymongo==3.11.2
PyMySQL==0.10.1
PySocks==1.7.1
pytz==2020.4
redis==3.5.3
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
wrapt==1.12.1
